fav - 2

+++ app/views.py
@@ -28,9 +28,11 @@ def getRecord(request):
     id = request.GET.get('id')
     record = get_record(id)
     is_fav = False
-    if len(FavRecord.objects.filter(user=request.user, recordId=id)) > 0:
+
+    if request.user.is_authenticated and len(FavRecord.objects.filter(user=request.user, recordId=id)) > 0:
         is_fav = True
     record['is_fav'] = is_fav
+    record['is_user_authenticated'] = request.user.is_authenticated
     return JsonResponse(record, safe=False)
 
 def getArtist(request):
@@ -42,6 +44,11 @@ def fav(request):
     if request.method == 'POST':
         id = request.POST.get('id')
         response = HttpResponse()
+
+        if not request.user.is_authenticated:
+            response.status_code = 403
+            return response
+
         FavRecord.objects.create(recordId=id, user=request.user)
         response.status_code = 200
         return response
@@ -52,6 +59,11 @@ def unfav(request):
     if request.method == 'POST':
         id = request.POST.get('id')
         response = HttpResponse()
+
+        if not request.user.is_authenticated:
+            response.status_code = 403
+            return response
+
         record = FavRecord.objects.filter(recordId=id, user=request.user)
         record.delete()
         response.status_code = 200
